refactor: 패키지 수정 및 Car 클래스 위치 이동
