+++ src/main/java/racingcar/Application.java
@@ -1,6 +1,6 @@
 package racingcar;
 
-import racingcar.Controller.GameController;
+import racingcar.controller.GameController;
 
 public class Application {
     public static void main(String[] args) {
